moved cloud tests
